componenet 추가
