+++ src/component/Navbar.module.css
@@ -1,0 +1,61 @@
+/* navbar 스타일 */
+.navbar {
+  background-color: #0D1B50;
+  color: #BBFD92;
+  font-weight: 600;
+}
+
+.conLarge {
+  width: 1440px;
+  margin: 0 auto;
+}
+
+.navbar .navCon{
+  height: 50px;
+  display:flex;
+  justify-content: space-between;
+  align-items: center;
+  /*background-color: red;*/
+}
+
+.navbar ul {
+  text-align: center;
+  display: flex;
+  justify-content: space-around;
+  font-size: 1rem;
+}
+
+.navbar .click{
+  flex-grow: 2;
+  margin-left: 150px;
+  margin-right: 130px;
+}
+
+.navbar .login{
+  flex-grow: 0.5;
+  margin-left: 30px;
+}
+
+.navbar .navSearch {
+  position: relative;
+  width: 150px;
+}
+
+.navbar .inputSearch {
+  width: 100%;
+  height: 30px;
+  padding-left: 20px;
+  font-size: 15px;
+  outline: none;
+  border: 0;
+  border-radius: 15px;
+  background-color: rgb(104, 111, 134);
+}
+
+.navbar .navSearch svg {
+  position: absolute;
+  right: 0%;
+  top: 50%;
+  transform: translateY(-50%);
+  color: #0D1B50;
+}
