+++ src/component/SelectDropBox.module.css
@@ -1,0 +1,19 @@
+/**/
+.selctWrap{
+  width: 200px;
+  height: 30px;
+  border : 1px solid rgb(37, 37, 37);
+  border-radius: 5px;
+  background-color: aqua;
+}
+
+select {
+  width: 100%;
+  height: 100%;
+  font-size: 15px;
+  text-align: center;
+  border:0; /*기본 스타일 제거 */
+  box-sizing: border-box;/*elect 박스의 크기 방식 지정*/
+  background:transparent;/*배경색 투명 처리*/
+
+}
