+++ src/component/TDcontent.js
@@ -2,12 +2,10 @@ import styles from "./TDcontent.module.css"
 
 const TDcontent = () => {
   return(
-    <>
       <div className={styles.TDcontentBox}>
+        <div className="hihi"></div>
       </div>
-    </>
   )
-
 }
 
 export default TDcontent;
